p2_worker/requirements.txt
boto3==1.40.25
python-dotenv==1.1.1
requests==2.31.0